move pkg/util/structs to test/util/structs
