+++ pkg/api/2019-04-30/types_test.go
@@ -10,8 +10,8 @@ import (
 	"github.com/Azure/go-autorest/autorest/to"
 
 	"github.com/openshift/openshift-azure/pkg/api"
-	"github.com/openshift/openshift-azure/pkg/util/structs"
 	"github.com/openshift/openshift-azure/test/util/populate"
+	"github.com/openshift/openshift-azure/test/util/structs"
 )
 
 var marshalled = []byte(`{
