+++ test/util/structs/walk.go
